+++ src/components/NavBar.vue
@@ -24,18 +24,11 @@
 
       <div class="navbar-end">
         <div class="navbar-item">
-          <button class="white--text" @click="exportProject">
-            EXPORT YOUR PROJECT
-          </button>
-          <span>
-            <!-- <v-icon class="save-icon" @click="saveProjectJSON">save_alt</v-icon> -->
-            <button class="save-icon" @click="saveProjectJSON">SAVE</button>
-          </span>
-          <button class="white--text" @click="openProjectJSON">OPEN</button>
-          <i class="fas fa-save fa-lg"></i>
-
-          <i class="fas fa-file-export fa-lg"></i>
-          <i class="fas fa-folder-plus fa-lg" @click="addProject"></i>
+          <SaveProjectComponent />
+          <OpenProjectComponent />
+          <NewProjectComponent />
+          <!-- <i class="fas fa-save fa-lg"></i> -->
+          <ExportProjectComponent />
         </div>
       </div>
     </div>
@@ -43,154 +36,21 @@
 </template>
 
 <script>
-import { mapState } from 'vuex';
-import fs from 'fs-extra';
-import path from 'path';
-import { addProject, changeTabName } from '../store/types';
-import localforage from 'localforage';
-
-const ipc = require('electron').ipcRenderer;
+// import { addProject } from '../store/types';
+// import localforage from 'localforage';
+import SaveProjectComponent from '@/components/SaveProjectComponent.vue';
+import OpenProjectComponent from '@/components/OpenProjectComponent.vue';
+import NewProjectComponent from '@/components/NewProjectComponent.vue';
+import ExportProjectComponent from '@/components/ExportProjectComponent.vue';
 
 export default {
   name: 'NavBar',
   props: ['route'],
-  methods: {
-    addProject() {
-      this.$store.dispatch(addProject, {
-        filename: 'untitled-' + this.$store.state.projectNumber,
-        lastSavedLocation: ''
-      });
-    },
-    exportProject: function() {
-      ipc.send('show-export-dialog');
-    },
-    saveProjectJSON() {
-      let projectLocation = this.$store.state.projects[
-        this.$store.state.activeTab
-      ].lastSavedLocation;
-      if (projectLocation) {
-        fs.writeFileSync(
-          projectLocation,
-          JSON.stringify(this.$store.state, null, 2)
-        );
-        console.log('PROJECT SAVED TO LAST SAVED LOCATION');
-      } else {
-        ipc.send('show-save-json-dialog');
-      }
-    },
-    openProjectJSON() {
-      ipc.send('show-open-dialog');
-    },
-    createComponentCode(componentLocation, componentName, children) {
-      fs.writeFileSync(
-        componentLocation + '.vue',
-        this.writeTemplate(children) +
-          this.writeScript(componentName, children) +
-          this.writeStyle(componentName)
-      );
-      console.log(children);
-    },
-    writeTemplate(children) {
-      let str = '';
-      children.forEach(name => {
-        str += `\t\t<${name}>\n\t\t</${name}>\n`;
-      });
-      return `<template>\n\t<div>\n${str}\t</div>\n</template>`;
-    },
-    writeScript(componentName, children) {
-      let str = '';
-      children.forEach(name => {
-        str +=
-          componentName === 'App'
-            ? `import ${name} from '@/components/${name}.vue';\n`
-            : `import ${name} from './${name}.vue';\n`;
-      });
-      let childrenComponentNames = '';
-      children.forEach(name => {
-        childrenComponentNames += `\t\t${name},\n`;
-      });
-      return `\n\n<script>\n${str}\nexport default {\n\tname: '${componentName}',\n\tcomponents: {\n${childrenComponentNames}\t}\n};\n<\/script>`;
-    },
-    writeStyle(componentName) {
-      let style =
-        componentName !== 'App'
-          ? ''
-          : `#app {\n\tfont-family: 'Avenir', Helvetica, Arial, sans-serif;\n\t-webkit-font-smoothing: antialiased;\n\t-moz-osx-font-smoothing: grayscale;\n\ttext-align: center;\n\tcolor: #2c3e50;\n\tmargin-top: 60px;\n}\n`;
-      return `\n\n<style scoped>\n${style}</style>`;
-    },
-    parseFileName(file) {
-      //'asdf/asdff/sdf.txt -> sdf.txt
-      return file.split('/').pop();
-    }
-  },
-  computed: {
-    ...mapState(['componentMap'])
-  },
-  mounted() {
-    ipc.on('export-project-location', (event, data) => {
-      if (!fs.existsSync(data)) {
-        fs.mkdirSync(data);
-        console.log('FOLDER CREATED!');
-      }
-      fs.copySync(
-        '/Users/Hubert/Desktop/Codesmith/ProductionProject/dev/PreVue/vue-boiler-plate',
-        data
-      );
-      for (let componentName in this.componentMap) {
-        if (componentName === 'App')
-          this.createComponentCode(
-            path.join(data, 'src', componentName),
-            componentName,
-            this.componentMap[componentName].children
-          );
-        else
-          this.createComponentCode(
-            path.join(data, 'src', 'components', componentName),
-            componentName,
-            this.componentMap[componentName].children
-          );
-      }
-    });
-    ipc.on('save-json-location', (event, data) => {
-      //delete original key from local forage
-      let deleteKey = this.$store.state.projects[this.$store.state.activeTab]
-        .filename;
-      localforage
-        .removeItem(deleteKey)
-        .then(function() {
-          console.log(deleteKey, 'Key is cleared!');
-        })
-        .catch(function(err) {
-          console.log(err);
-        });
-
-      let fileName = this.parseFileName(data);
-      // this.$store.dispatch(changeTabName, fileName);
-      this.$set(this.$store.state.projects, this.$store.state.activeTab, {
-        filename: fileName,
-        lastSavedLocation: data
-      });
-      // console.log('DATA[0]', data);
-      fs.writeFileSync(data, JSON.stringify(this.$store.state, null, 2));
-      localforage
-        .setItem(fileName, JSON.parse(fs.readFileSync(data, 'utf8')))
-        .then(() => console.log('saved ', fileName, 'to local forage'));
-
-      console.log('PROJECT SAVED AS A JSON OBJECT!');
-    });
-    ipc.on('open-json-location', (event, data) => {
-      //set state of local forage
-      let fileName = this.parseFileName(data[0]);
-      localforage
-        .setItem(fileName, JSON.parse(fs.readFileSync(data[0], 'utf8')))
-        .then(() => console.log('saved ', fileName, 'to local forage'));
-      //create new tab name with file name
-      this.$store.dispatch(addProject, {
-        filename: fileName,
-        lastSavedLocation: data[0]
-      });
-      //when tab is switch, it will go thr right path.
-    });
+  components: {
+    SaveProjectComponent,
+    OpenProjectComponent,
+    NewProjectComponent,
+    ExportProjectComponent
   }
 };
 </script>
